Fixed WithLimit option and added test
